Add redeploy webhook
Added another GitHub webhook to re-deplopy the application following a
GitHub push.

This is different from the existing gitub-webhook.js application in that
it is triggered by a push to the pattern library UI repo, rather than
the pattern library repo. It also does a git pull on this repo to get
the latest UI elements before running a rebuild.

@todo Combine these two webhooks to make this more elegant